+++ lib/src/virtual_renderer.dart
@@ -3,11 +3,17 @@ part of ace;
 /// The renderer draws to the screen.
 abstract class VirtualRenderer extends _Disposable implements OptionsProvider {
   
+  /// The root element containing this renderer.
+  html.Element get containerElement;
+  
   /// The index of the first visible row.
   int get firstVisibleRow;
   
   /// Whether or not the gutter has a fixed width.
   bool fixedWidthGutter;
+  
+  /// The element that the mouse events handlers are attached to.
+  html.Element get mouseEventTarget;
   
   /// The column number of where the print margin is.
   int printMarginColumn;